+++ messages/src/did_doc/aries/mod.rs
@@ -1,2 +1,1 @@
 pub mod diddoc;
-pub mod service;
