+++ messages/src/protocols/out_of_band/service_oob.rs
@@ -1,4 +1,4 @@
-use crate::did_doc::aries::service::AriesService;
+use crate::concepts::aries_service::AriesService;
 use crate::protocols::connection::did::Did;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
